Moved some shared parameters back to the caller module

+++ test.py
@@ -36,6 +36,10 @@ class Test:
         if not os.path.exists(self.train_dir):
             raise FileNotFoundError('Could not find folder {}'.format(self.train_dir))
         if os.path.exists(self.test_dir):
+            eprint('Confirm removing {}\n[Y/n]'.format(self.test_dir))
+            if input() != 'Y':
+                import sys
+                sys.exit()
             import shutil
             shutil.rmtree(self.test_dir)
             eprint('Removed: ' + self.test_dir)
@@ -187,20 +191,25 @@ def main(argv=None):
     argp = argparse.ArgumentParser()
     # testing parameters
     argp.add_argument('dataset')
-    argp.add_argument('--random_seed', type=int)
+    argp.add_argument('--random-seed', type=int)
     argp.add_argument('--device', default='/gpu:0')
     argp.add_argument('--postfix', default='')
-    argp.add_argument('--train_dir', default='./train{postfix}.tmp')
-    argp.add_argument('--test_dir', default='./test{postfix}.tmp')
+    argp.add_argument('--train-dir', default='./train{postfix}.tmp')
+    argp.add_argument('--test-dir', default='./test{postfix}.tmp')
     argp.add_argument('--log-file', default='test.log')
     argp.add_argument('--batch-size', type=int, default=1)
     # data parameters
+    argp.add_argument('--dtype', type=int, default=2)
+    argp.add_argument('--data-format', default='NCHW')
     argp.add_argument('--patch-height', type=int, default=512)
     argp.add_argument('--patch-width', type=int, default=512)
+    argp.add_argument('--in-channels', type=int, default=3)
+    argp.add_argument('--out-channels', type=int, default=3)
     # pre-processing parameters
     input_arguments(argp)
     # model parameters
     SRN.add_arguments(argp)
+    argp.add_argument('--scaling', type=int, default=1)
     # parse
     args = argp.parse_args(argv)
     args.train_dir = args.train_dir.format(postfix=args.postfix)
